Add deprecated environment variable scanning
This commit adds a new feature to scan for deprecated environment variables in the code. It introduces a new module `hemera.deprecated` that contains the logic for scanning and logging warnings for deprecated environment variables. The `scan_for_deprecated_env_vars` function is called in the main function of `github/__init__.py` to check for deprecated environment variables and raise an error if necessary. This feature helps developers migrate from deprecated environment variables to their replacements.